cogact: cast timestep embedding to model weight dtype
timestep_embedding() always creates float32 tensors; when model.to(bfloat16)
is called, the MLP weights become bfloat16 but t_freq stays float32, causing
RuntimeError: mat1 and mat2 must have the same dtype.

+++ cogact/pytorch/src/model.py
@@ -43,7 +43,7 @@ class TimestepEmbedder(nn.Module):
 
     def forward(self, t):
         t_freq = self.timestep_embedding(t, self.frequency_embedding_size)
-        return self.mlp(t_freq)
+        return self.mlp(t_freq.to(self.mlp[0].weight.dtype))
 
 
 class LabelEmbedder(nn.Module):
